// store/ideaNft/mutation-types.ts
import { ActionTypes } from "xstate";

export enum MutationTypes {

    SET_EMAIL = "SET_EMAIL",
    SET_LOGGED_IN = "SET_LOGGED_IN",

    // TOKEN
    GET_ACCESS_TOKEN = "GET_ACCESS_TOKEN",
    SET_ACCESS_TOKEN = "SET_ACCESS_TOKEN",
    SET_REFRESH_TOKEN = "SET_REFRESH_TOKEN",

    // USER DETAILS
    SET_USER_ROLES = "SET_USER_ROLES",
    SET_FIRST_NAME = "SET_REFRESH_TOKEN",
    SET_LAST_NAME = "SET_LAST_NAME",
    SET_AVATAR_URL = "SET_AVATAR_URL",
    SET_USER_INPUT_ERROR = "SET_USER_INPUT_ERROR",

    SET_IDEA_NFT = "SET_IDEA_NFT",
    SET_OWNERSHIP_ROYALTY_FOR_IDEA_NFT = "SET_OWNERSHIP_ROYALTY_FOR_IDEA_NFT",

    SET_ACTIVITY_IDEA_NFT = "SET_ACTIVITY_IDEA_NFT",
    SET_INVITATIONS_IDEA_NFT = "SET_INVITATIONS_IDEA_NFT",

    SET_IDEA_NFT_LISTED = "SET_IDEA_NFT_LISTED",

    SET_ONBOARDING_STATES = "SET_ONBOARDING_STATES",
    SET_ACTION_TYPES = "SET_ACTION_TYPES",
    SET_ACTIVITY_TYPES = "SET_ACTIVITY_TYPES",
    SET_IDEA_STATES = "SET_IDEA_STATES",
    SET_ENCRYPTION_KEY = "SET_ENCRYPTION_KEY"
}